requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.28.3
psutil==6.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.17.0
setproctitle==1.3.3
setuptools==75.2.0
six==1.16.0
smmap==5.0.1
sympy==1.13.1
torch==2.5.0
torchvision==0.20.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Wand==0.6.13
wandb==0.18.5
